+++ MWSTAT.API.Domain/Dtos/InstagramRunDetailsToInstagramUserDto.cs
@@ -7,7 +7,7 @@ namespace MWStat.API.Domain.Dtos
         public int Id { get; set; }
         public int InstagramRunDetailsId { get; set; }
         public virtual InstagramRunDetailsDto InstagramRunDetails { get; set; }
-        public int InstagramUserId { get; set; }
+        public long InstagramUserId { get; set; }
         public virtual InstagramUserDto InstagramUser { get; set; }
         public RelationToUserEnum RelationToUser { get; set; }
     }
