+++ MWSTAT.API.Domain/Dtos/InstagramSessionDataDto.cs
@@ -3,7 +3,7 @@
     public class InstagramSessionDataDto
     {
         public int Id { get; set; }
-        public int InstagramUserId { get; set; }
+        public long InstagramUserId { get; set; }
         public virtual InstagramUserDto InstagramUser { get; set; }
         public string AuthToken { get; set; }
         public string SessionData { get; set; }
