+++ MWSTAT.API.Domain/Models/InstagramRunDetailsToInstagramUser.cs
@@ -7,7 +7,7 @@ namespace MWStat.API.Domain.Models
         public int Id { get; set; }
         public int InstagramRunDetailsId { get; set; }
         public InstagramRunDetails InstagramRunDetails { get; set; }
-        public int InstagramUserId { get; set; }
+        public long InstagramUserId { get; set; }
         public InstagramUser InstagramUser { get; set; }
         public RelationToUserEnum RelationToUser { get; set; }
     }
